requirements.txt
alembic==1.12.0
amqp==5.2.0
aniso8601==9.0.1
billiard==4.2.0
blinker==1.6.2
cachetools==5.3.3
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.38.0
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
limits==3.7.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
pandas==2.2.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycountry==23.12.11
Pygments==2.17.2
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-slugify==8.0.3
python-telegram-bot==21.3
pytz==2023.4
qrcode==7.4.2
redis==5.0.1
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
six==1.16.0
SQLAlchemy==2.0.25
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2